vectorstore_service/app/requirements.txt
flask
faiss-cpu
numpy
gunicorn
